PL-16 checkstyle

+++ ThePlugin/src/main/java/com/atlassian/theplugin/idea/bamboo/table/columns/BuildTestRatioColumn.java
@@ -41,7 +41,7 @@ public class BuildTestRatioColumn extends BambooColumnInfo {
 	private double getTestRatio(Object o) {
 		double oTests = 0;
 		if (((BambooBuildAdapter) o).getTestsNumber() > 0) {
-			oTests = ((BambooBuildAdapter) o).getTestsPassed() / ((BambooBuildAdapter) o).getTestsNumber();
+			oTests = (double) (((BambooBuildAdapter) o).getTestsPassed() / ((BambooBuildAdapter) o).getTestsNumber());
 		}
 		return oTests;
 	}
